refactor: models folder clinics\models
